+++ example/lib/sign_up.dart
@@ -45,30 +45,25 @@ class SignUp extends StatelessWidget {
           const Divider(),
           optionText,
           spacer,
-          signInBtn(context, Icons.email, 'Sign in with Magic Link', () {
-            Navigator.popAndPushNamed(context, '/magic_link');
-          }),
+          ElevatedButton.icon(
+            onPressed: () {
+              Navigator.popAndPushNamed(context, '/magic_link');
+            },
+            icon: const Icon(Icons.email),
+            label: const Text('Sign in with Magic Link'),
+          ),
           spacer,
-          signInBtn(context, Icons.phone, 'Sign up with Phone', () {
-            Navigator.popAndPushNamed(context, '/phone_sign_up');
-          }),
+          ElevatedButton.icon(
+            onPressed: () {
+              Navigator.popAndPushNamed(context, '/phone_sign_up');
+            },
+            icon: const Icon(Icons.phone),
+            label: const Text('Sign up with Phone'),
+          ),
           spacer,
           SupaSocialsAuth(
             colored: true,
-            socialProviders: const [
-              SocialProviders.apple,
-              SocialProviders.azure,
-              SocialProviders.bitbucket,
-              SocialProviders.discord,
-              SocialProviders.facebook,
-              SocialProviders.github,
-              SocialProviders.gitlab,
-              SocialProviders.google,
-              SocialProviders.slack,
-              SocialProviders.spotify,
-              SocialProviders.twitch,
-              SocialProviders.twitter,
-            ],
+            socialProviders: SocialProviders.values,
             onSuccess: (session) {
               Navigator.of(context).pushReplacementNamed('/home');
             },